Package into escripts for release
I had been struggling with how to package these applications for release. A typical escript will embed Elixir, which is not good for us because the user will want to use his system-wide Elixir installation.

We now create an escript for language_server and debugger but we don't embed elixir. Instead, there's a shell script `exscript.sh` that attempts to set `ERL_LIBS` to include the system-wide Elixir when running the escripts.

TODO: Windows batch script.