Fix bug causing Dialyzer warnings to remain once fixed
The Dialyzer server was storing warnings in a map where the key is the path for the beam file with the warning. If compilation is happening concurrently with the analysis, though, the module might have been unloaded and deleted, which causes Dialyzer to be unable to determine the correct path for the beam file. The warning is then stored with the key "non_existing" and is not overridden when the beam file is re-analyzed.

This fixes it by changing the warnings map key to use the module itself instead of the beam file path.