Hack: Override MIX_EXS in launcher scripts so that Mix won't load mixfile in cwd on launch
Mix will automatically load the mixfile at the current directory which is unsafe because it may print to stdout before we've begun intercepting output. The launch scripts now change the environment variable `MIX_EXS` so that won't happen, and then we restore it from `ELIXIR_LS_MIX_EXS` on the slim chance the user actually has set `MIX_EXS` to something.