Simplify `with_build_lock`
It's safer to lock using the PID of the caller, otherwise there's a risk that two processes can acquire the same lock. Also renamed the lock to __MODULE__ to decrease likelihood of a name collision somewhere (though it's unlikely). Since `with_build_lock` no longer requires the requesting module as a parameter, I removed the logging output as well.