Path file path instead of contents to :dialyzer_utils.get_core_from_beam/1
Under the hood, this function can take either a charlist with the file path or the file's contents as a binary. However, if it fails (due to no debug_info chunk, for example), it will crash when the argument is the file's contents instead of path. Unfortunately, changing this means we do extraneous file reads from disk, but it's worth it for reliability.