Fixes a bug in dependency loading in Elixir 1.9 and fixes bug causing builds to fail permanently after an invalid dep specification
Mix.Dep.loaded/1 was removed in Elixir 1.9. This was also a problem in the mix_task_archive_deps dependency we use in the release task, so I've updated a fork of that repo and am sending a PR upstream.

Also, checking cached deps to look for changed deps can fail, so we now rescue those failures.

Fixes #155