Bump elixir sense (#42)
* Bump elixir sense

Fixes #30 and brings in a host of other improvements/enhancements
(although we need some updates to make full use of the elixir_sense
improvements)

* Bump elixir_sense again to get fix for encountered issue

https://github.com/elixir-lsp/elixir_sense/pull/38

* Bump elixir sense and fix tests

Encountered an ElixirSense bug along the way that was filed as:
https://github.com/elixir-lsp/elixir_sense/issues/42

And necessitated changing the references text fixture

* Add a basic travis config

* Give the server more time to respond in tests

Also refactor the assertion to first find the response to the message and then
check for format. This makes it much easier to debug the output of a failing
test, especially to tell the difference between no response with that id and an
incorrect response with that id.