Test each dialyzer test format
Also fix the reporting of diagnostics by using `Task.start_link/1` instead of
`spawn/1`. The issue with spawn is that it would start the process outside of
the supervision tree. So when we called `GenServer.stop/1` in the test the
`ElixirLS.LanguageServer.Server` would stop, which would bring down the
`ElixirLS.LanguageServer.Dialyzer` GenServer, but the
`ElixirLS.LanguageServer.Dialyzer.Manifest` process would keep running. However,
the `ElixirLS.LanguageServer.Dialyzer.Manifest` process depends on an ETS table
owned by `ElixirLS.LanguageServer.Dialyzer`, so the ETS calls would start to
fail. Using `Task.start_link/1` fixes this because it causes the task to die at
the same time that `ElixirLS.LanguageServer.Dialyzer` dies, which means that no
process tries to access a non-existant ETS table.