Default to dialyxir_long dialyzer format (#45)
* Default to dialyxir_long dialyzer format

* Test each dialyzer test format

Also fix the reporting of diagnostics by using `Task.start_link/1` instead of
`spawn/1`. The issue with spawn is that it would start the process outside of
the supervision tree. So when we called `GenServer.stop/1` in the test the
`ElixirLS.LanguageServer.Server` would stop, which would bring down the
`ElixirLS.LanguageServer.Dialyzer` GenServer, but the
`ElixirLS.LanguageServer.Dialyzer.Manifest` process would keep running. However,
the `ElixirLS.LanguageServer.Dialyzer.Manifest` process depends on an ETS table
owned by `ElixirLS.LanguageServer.Dialyzer`, so the ETS calls would start to
fail. Using `Task.start_link/1` fixes this because it causes the task to die at
the same time that `ElixirLS.LanguageServer.Dialyzer` dies, which means that no
process tries to access a non-existant ETS table.

* Fix dialyzer format test on Elixir 1.9

* Bump OTP minimum version to 20

This is because the dialyzer implementation relies on
`dialyzer_utils:get_core_from_beam/1` which was added in OTP 20 along with the
Dbgi chunk in:

https://github.com/erlang/otp/commit/dfb899c0229f7ff7dbfad34d496e0429562728bf