Fix Manifest race condition (#72)
The anonymous function spawned in `ElixirLS.LanguageServer.Dialyzer.Manifest`
was not guaranteed to stop before it's parent (`ElixirLS.LanguageServer.Server`).
In the tests this would manifest itself as an occasion error when the Manifest
spawned process would try to read from an ETS table owned by the server after
the server was already dead which resulted in an `ArgumentError`. An example of
this error occurring can be found at:
https://travis-ci.org/elixir-lsp/elixir-ls/jobs/612862401

By changing the `spawn_link` to a `Task.start_link` we ensure that the spawned
process (now a task) is properly supervised and is brought down before its
parent is, thus eliminating the race condition with the ETS table ownership.