Bring textDocumentSync save option up to spec
The current Language Server Protocol specification (3.15 at the time of
writing) dictates that the `save` value  of `TextDocumentSyncOptions`
should, if defined, be of type `SaveOptions`, which it in turn defines
as an object with an optional `includeText` property of type `boolean`.

This changes the `textDocumentSync["save"]` value, which was set to a
bare boolean, to a map with the expected `includeText` boolean value.