DocumentSymbols provider should handle modules without a name
The DocumentSymbols provider should be able to handle a variety of compilation
errors so that it can be useful while debugging the compilation errors. The main
problem was that the `extract_symbol/2` clause that handled `defmodule` and
`defprotocol` was overly restrictive in the pattern match that it did. So I've
changed it to match all the possible AST variations (although some may still
raise an exception within the clause, but this prevents bad data from
propagating).

`MISSING_MODULE_NAME` was returned since a non-empty name is required