DocumentSymbols provider handles compilation errors gracefully
Rather than throwing an exception (which causes the spawned process to crash)
the DocumentSymbols provider will now simply return an error.

This is beneficial because this is very much an expected error and thus it is
confusing to show the exception in the logs. Possibly there are also performance
benefits to this as well.

Ideally the DocumentSymbols provider could still return valid symbols in this
case, but that is currently infeasible.