Fix dialyzer test when running locally
I'm testing running locally on arch linux. It appear that for me locally
for the "reports diagnostics then clears them once problems are fixed"
test I get the output I get the message:

    [ElixirLS Dialyzer] Analyzing 3 modules: [A, B, :erlang]

Instead of:

    [ElixirLS Dialyzer] Analyzing 2 modules: [A, B]

Adding a function definition in module C causes :erlang to not be
detected as needing to be analyzed.