Further improve the tests
The previous `:elixir_ls_test_process` approach was not reliable and seemed to
suffer from a race condition. Instead always manually start the server so we can
be in full control of the processes.

Unfortunately aliases are not well supported in umbrella projects so use
`cmd mix test` as specified in:

https://elixirforum.com/t/mix-task-aliases-and-umbrellas/12666