Remove dialyzer supervisor
Fixes #101

I don't see why `ElixirLS.LanguageServer.Dialyzer` should have a supervisor, but
all the other GenServers do not. They should all be treated the same.

I'm not completely sold on this approach. Instead it would probably be better to
have a solid supervision hierarchy instead of relying on client to restart the
server. For VSCode that happens automatically, but for other clients it does
not, and we lose state that the server is relying on (such as the currently open
files)