Add +swbt none to remove busy-wait and improve performance
Utilizing busy-wait to reduce latency primarily makes sense for
applications that are running as the only node on the machine. So it
doesn't make sense for ElixirLS because there will usually be other
processes running on the machine such as the editor and potentially
other ElixirLS instances.

A blog post that touches on `+swbt none`:
https://www.ably.io/blog/beam-optimization-mqtt/

Also add sbwtdcpu and sbwtdio

Fixes #328