Fix textDocument/formatting returns [] when format Ecto migration script
When try to format Ecto migration script that generate project from
Phoenix, it will returns inputs "*.exs". But the glob list just detect
it only the root of project and apps directory which for umbrella
project. This cause make migration script won't match any glob pattern.

This changes fixes by add directory of file that want to format to the end
of glob list.

Fixes #361