Refactor implementation of SourceFile.lines_with_endings
The previous implementation did not have any bugs (that I saw), but I
did find it more difficult to read and understand. I expect performance
to be comparable.

For more comprehensive testing I have added StreamData which adds
property-based testing. We can probably use that in other portions of
the code base as well.