Make wrapper script more robust... more robustly
This makes a setup possible in which a symlink points to the file
(e.g. /usr/local/bin/elixir-ls -> /opt/elixir-ls/language_server.sh).

It's a rehash of a previous commit that was reverted since it broke
setups where the language server was installed in a path containing
spaces (e.g. /opt/Elixir Language Server/). This variant takes that
possibility into account.