Make wrapper script more robust... more robustly (#473)
* Make wrapper script more robust... more robustly

This makes a setup possible in which a symlink points to the file
(e.g. /usr/local/bin/elixir-ls -> /opt/elixir-ls/language_server.sh).

It's a rehash of a previous commit that was reverted since it broke
setups where the language server was installed in a path containing
spaces (e.g. /opt/Elixir Language Server/). This variant takes that
possibility into account.

* Avoid undefined behaviour in launch.sh

I'm not really sure what could lead this to happen. Maybe some weird
directory permissions with the read bit on and the exec bit off...

Still good to be careful though.

Co-authored-by: Łukasz Samson <lukaszsamson@gmail.com>

Co-authored-by: Joshua Trees <me@jtrees.io>
Co-authored-by: Łukasz Samson <lukaszsamson@gmail.com>