Support describes and tests with unevaluated names (#537)
When a test is formatted as:

    test "performs some" <> "action" do
      assert work_done()
    end

The language server crashes with message:

    ** (Protocol.UndefinedError) protocol String.Chars not implemented for {:<>, ... ["performs some", "action"], ...

This comes from `lib/language_server/providers/document_symbols.ex:222`
where the "name" of the test is interpolated into a string. In order
for this "unevaluated" test name AST to be interpolated into a string,
it must first become a string or something that the String.Chars
protocol can turn into a string.

My approach here is to turn the test name AST into a string using
`Macro.to_string`, which simply converts an AST into its Elixir code
equivalent.