Handle more cases for manipulating pipes
Changed the function call detection to not "jump forward", as part of
this if the user requests an index that is not on a function call, then
no action is performed.

Also fix handling of piping function calls that span multiple lines by
adjusting the returned line and column to point to the correct start
location by counting backward.