Allow configuring debugExpressionTimeoutMs
Rather than a hard-coded timeout of 1_000 ms for watch expressions, add
`debugExpressionTimeoutMs` to the launch.json configuration options. In
the next version of vscode-elixir-ls `debugExpressionTimeoutMs` will be
supported in the launch.json autocomplete, but if you are running a
version of ElixirLS that supports it before then you can add the key
manually.

Example mix test launch.json entry that uses this to set a timeout of 30
seconds:
```
{
    "type": "mix_task",
    "name": "mix test",
    "request": "launch",
    "debugExpressionTimeoutMs": 30000,
    "task": "test",
    "taskArgs": [
        "--trace"
    ],
    "startApps": true,
    "projectDir": "${workspaceRoot}",
    "requireFiles": [
        "test/**/test_helper.exs",
        "test/**/*_test.exs"
    ]
}
```

Fixes #525