Fix formatting performance problems with .formatter.exs in subdirectories (#609)
* Convert should_format? test to use format/3 and fixtures

This function is really an implementation detail, so let’s test the
public API.

The function also relies on interaction with the filesystem, so testing
without fixtures is a bad idea because it will lead to incorrect
results.

* Check .formatter.exs :inputs relative to its directory

When dealing with `.formatter.exs` files that are in subdirectories of
the main project, the behaviour was incorrect:
https://github.com/elixir-lsp/elixir-ls/issues/361

And also caused performance problems:
https://github.com/elixir-lsp/elixir-ls/issues/381

A good example is a project I currently work on. We have a top-level
`.formatter.exs`, as well as a `test/.formatter.exs`.

The one in `test/` has the `:inputs` option set to `["**/*.{ex,exs}"]`.
Before this commit, should_format?/3 would take this pattern and
concatenate it with the project directory. This would result in
Path.wildcard/2 globbing for all source files in the entire project,
rather than just under test/. This causes formatting to take about 5
seconds on our project.

This commit fixes the problem by taking the file to be formatted, and
finding the `.formatter.exs` file that exists in the closest parent
directory (so long as that directory is still within the project root).
We then perform the Path.wildcard/2 relative to that directory.

I’ve removed the hack that was in place to support umbrella projects,
because this is a more general solution to the problem.

On our project, I am now able to format a file in 100ms or so, which
isn’t a bad speed-up!

* Avoid costly Path.wildcard/2 call

The call to Path.wildcard/2 can cause use to traverse a large number of
files, which can be very slow depending on the project and the globs
used:

https://github.com/elixir-lsp/elixir-ls/issues/381

I wrote a library that implements a glob parser and allows us to check
whether a glob matches a path without touching the filesystem:

https://github.com/jonleighton/path_glob

This commit switches out the Path.wildcard/2 call to use this library.

Co-authored-by: Jason Axelson <axelson@users.noreply.github.com>