Re-use document symbols functionality for workspace symbols
Previously, the workspace symbols feature used the `:code.all_loaded`
functionality of the OTP standard library to "discover" all modules, and
then was able to parse out the different type of symbols from that.

In doing so, this loaded _tons_ of symbols that are completely unrelated
to your workspace. As a prior user of the document symbols feature, this
really threw me off. As I understand it, the workspace symbols support
is for quick navigation of functions, modules, types in your own
codebase, so if the source code is not available in your own repository,
you really wouldn't be navigating to it.

The workspace symbols feature was also implemented differently than the
document symbols feature. This commit works to align the two features to
make them work similarly.

In addition, this also allows you to pass an empty query to LSP request
in order to retrieve a full list of all of the symbols. This aligns the
Elixir implementation with the spec, and makes it useful for those who
use a fuzzy searching plugin in their editor to do the querying.

_Notes_

- I believe this is technically a breaking change, as the results
  returned in the request are slightly different than they were before.
  I believe them to be _more_ accurate now, but they are technically
  different.
- With the previous implementation, I found that in some codebases (like
  https://github.com/mhanberg/temple), no symbols were being returned
  for the project. I suspect it had something to do with the library not
  having an `Application`, so no code is loaded at first.
- I have tested this on the Wallaby codebase
  (https://github.com/elixir-wallaby/wallaby) and it seems to return
  instantly, so I don't think there should be any kind of performance
  regression, but it would be worth checking against larger codebases. I
  plan on trying this branch on my work projects on my next working day.