Special case handling of Phoenix 'live' dir when suggesting module names (#705)
When generating a live view with mix phx.gen.live the view and all related files
are put under my_project_web/live/*

Similar to other Phoenix paths (e.g. /controllers) this directory is
not a part of the generated module names, and the suggested module names
when auto-completing should account for that convention.