Fixed language server crashes
While editing in emacs, I've been noticing a lot of language server
crashes. It appears that some of them are due to the way emacs builds
elixir files, which can cause the tooling of elixir-ls to emit syntax
errors or have off-by-one line errors or not handle certain types of
code.

I might be barking up the wrong tree here, but I think elixir-ls
should be pretty agressive with handling errors and shouldn't crash,
but if that's the expected behavior, then we'll need a different
approach, and I'm all ears.

The changes here reduce the number of crashes dramatically so it's
pleasant to use elixir-ls in emacs again.