Improved resiliency of `formatter_for`
It's possible that the call to `Mix.Tasks.Format.formatter_for_file`
can cause a compile or syntax errors. Because these types of errors
don't have a `message` key, the attempt to log them will fail, and the
server will crash.

This PR uses `Exception.message` to build the message text, which
should work on any exceptions.