SourceFile experimental replacement
This PR represents a suggestion on how elixir-ls can evolve the concept
of the source file. The current implementation is a group of
semi-related functions that operate on many different things, leading
to logic that's spread around the codebase, with each user of SourceFile
treating things differently.

This module, however always operates on the data structure it defines,
and provides a consistent and simplified interface for its users.
Among other things, it:

  * Builds changes with iolists, and doesn't need to do any reversing
    options
  * Introduces a data structure representing a document, with 0(1)
    random access.
  * Introduces a line record that separates the ending from the text
  * Introduces a new line parser that is significantly faster and more
    efficient than SourceFile.lines_with_endings
  * Applies edits in a single loop over the document