Fixed unit tests that were broken before 1.13
The unit tests on builds prior to 1.13 were failing because I was
patching the wrong function. Prior to 1.13, formatter_for_file didn't
exist, but the unit tests would patch that but the
`function_exported?` check in `source_file.ex` would skip that branch
and succeed, which would cause the tests to fail.

The change is simple, check to see if `formatter_for_file` exists, and
if it doesn't, patch the `formatter_opts_for_file` function