Ported ASCII detection / conversion from protocol branch
When parsing a file, we check if each line is ascii or not. If it is,
we can skip the conversions from utf-8 -> utf-16 -> utf8 when we apply
changes from the client. This means we reduce a ton of allocations and
simplify our code.

I also made a conversions module and some types representing Positions
and Ranges. Don't read too much into them, because I have a more
extensive data model that I'd like to discuss.