Removed utf16 conversions during document changes
Prior to this change, document tracking required doing utf-8 to utf-16
conversion on every change. Now all conversion is done on the
positions that are sent to us. We still need to do conversions on
non-ascii lines, but this should greatly cut down on the number of
conversions that elixir-ls has to do.

There was also an off-by-one error in the `conversions.ex` file where
we looked up the wrong line to do conversions. When utf16 conversions
were removed from `source_file.ex`, the bug became apparent with
several unit tests failing.

This also includes a small change to heuristically detect utf16 text.