Change how requests and notfications are structured
There are a number of conversions related to document positions and
document realization that need to happen when a request is
serviced. Future request handlers would be made much simpler and more
correct if the system would handle these conversions for writers of
provider handlers. To wit, each request and notification module now
has an internal `lsp` field that contains the initial request, while
the surounding request represents the elixir side of the data. To
convert to elixir, all one has to do is use the `to_elixir` function.

The elixir struct that encloses the lsp struct has all of the same
fields, and if the enclosing struct has a text document, an additional
`source_file` field is added. All the fields on the elixir side are
initialized to nil until `to_elixir` is called.

When to_elixir is invoked, it scans the lsp document for
text_documents, which are fetched from the source file store. This
document is used to influence any ranges or positions and convert them
into their elixir-native data structures. Finally, all of the other
fields are copied into the elixir side.