Scohen/encapsulate source file (#743)
* SourceFile experimental replacement

This PR represents a suggestion on how elixir-ls can evolve the concept
of the source file. The current implementation is a group of
semi-related functions that operate on many different things, leading
to logic that's spread around the codebase, with each user of SourceFile
treating things differently.

This module, however always operates on the data structure it defines,
and provides a consistent and simplified interface for its users.
Among other things, it:

  * Builds changes with iolists, and doesn't need to do any reversing
    options
  * Introduces a data structure representing a document, with 0(1)
    random access.
  * Introduces a line record that separates the ending from the text
  * Introduces a new line parser that is significantly faster and more
    efficient than SourceFile.lines_with_endings
  * Applies edits in a single loop over the document

* Added benchmarks and line parser test

* Optimized slice, removed dead code

* experimental; changed text to to_string

* Made an experimental top-level package

This will help isolate the experimental stuff from the mainline more
cleanly, and also allows swapping of implementations via aliasing

* Added benchmark for document access

While playing around with the benchmark, I noticed that we weren't
consolidating protocols. This had a dramatic effect on the fetch
performance of the fetch_line call.

* Converted document backing store to tuples

* Added benchmark to test the speeds of various backing stores for Documents

* Ported ASCII detection / conversion from protocol branch

When parsing a file, we check if each line is ascii or not. If it is,
we can skip the conversions from utf-8 -> utf-16 -> utf8 when we apply
changes from the client. This means we reduce a ton of allocations and
simplify our code.

I also made a conversions module and some types representing Positions
and Ranges. Don't read too much into them, because I have a more
extensive data model that I'd like to discuss.

* Re-added Enum.at

* Fixed package name

* Performance improvements

* Changed document implementation to use tuples
* Enabled protocol consolidation outside of tests
* Change LineParser to use an indexed based approach

* Ported more improvements from the protocol branch

* Slightly simplified line parsing code

* Fixed utf conversion bugs

LineParser was not correctly setting the ascii? flag.
Also added an optimization to SourceFile to not convert the prefixes
and suffixes to utf16 if the lines are ascii.

* Removed utf16 conversions during document changes

Prior to this change, document tracking required doing utf-8 to utf-16
conversion on every change. Now all conversion is done on the
positions that are sent to us. We still need to do conversions on
non-ascii lines, but this should greatly cut down on the number of
conversions that elixir-ls has to do.

There was also an off-by-one error in the `conversions.ex` file where
we looked up the wrong line to do conversions. When utf16 conversions
were removed from `source_file.ex`, the bug became apparent with
several unit tests failing.

This also includes a small change to heuristically detect utf16 text.

* Removed utf16 conversions during document changes

Prior to this change, document tracking required doing utf-8 to utf-16
conversion on every change. Now all conversion is done on the
positions that are sent to us. We still need to do conversions on
non-ascii lines, but this should greatly cut down on the number of
conversions that elixir-ls has to do.

There was also an off-by-one error in the `conversions.ex` file where
we looked up the wrong line to do conversions. When utf16 conversions
were removed from `source_file.ex`, the bug became apparent with
several unit tests failing.

This also includes a small change to heuristically detect utf16 text.

* Removed protocol type

* Simplified prefix/suffix functions

* fixed import

* Fixed benchamrk off-by-one error

* Complied with out-of-spec unit tests

Some of the microsoft node tests deal with out of spec-lsp
messages. Prior, I was returning an error if an out-of-spec position
was encountered, but this change allows us to accept these messages
and handle them like the node client does.

* fixed merge error

* Revert "Complied with out-of-spec unit tests"

This reverts commit e07e0af8bd10b03d1a4867fdf87d563f2529be70.

* Added additional tests