Added code generators for LSP data structures
Adding the LSP data structures by hand was growing tedious and there
were a couple bugs. Doing this programatically seems like it would be
safer and easier in the long run.
Right now, all of the data structures are mapped, which will likely make
compile times a bit longer, but mapping things as needed would add a lot
of complexity to something that only needs to be done occasionally.

Note: as of now, I haven't run this and checked in any artifacts. That
will happen in a subsequent PR.