Made .formatter.exs inputs relocatable
The .formatter.exs defined paths via a wildcard, but this would only
work if it was executed from the current directory. If it was run from
the topmost directory, none of the paths would be found, and the
inputs would only contain the `.exs` files in the current directory.

I also removed config from the glob, as this is in an umbrella
project, and umbrella apps have a centralized config which isn't in
this directory.