Fix directory issue with `fish` launch script
Fish doesn't have subshells, so `cd`ing within a command substitution
will change the current working directory. When a directory-based
version manager like `asdf` or `rtx` is used, the change in cwd could
cause a different version of Elixir to run than that of the project.
This fixes that potential issue by saving the current directory before
calling the `readlink_f` function.