fix: remove assumption to fix deterministic builds
Current code has an assumption that `[:compile][:source]` will always
exist in the output of `Enum.module_info()`. This is stripped from
Elixir's deterministic build toggle
(`ERL_COMPILER_OPTIONS=deterministic`).

Added an assumption for `~"/"` as a fallback when such a case happens to
keep the old logging behaviour. This can be changed as seen fit.

Manual testing shows that elixir-ls will no longer throw on startup due
to `nil` given as a path with a deterministic build of Elixir.

Fixes #1197